Make screen_line_address the width it actually needs
It was long, but it only ever holds SCREEN_ADDRESS ($B800) plus at most
25*80 bytes -- a maximum of $BFD0, which is 16 bits.  The wide type
bought nothing: lpoke() and lcopy() take 28-bit addresses and widen it
for free, while screen_hex() and screen_decimal() take unsigned int and
had to narrow it.

That narrowing is what -Wconversion was reporting at
fdisk_hal_mega65.c:294, where the previous commit silenced it with a
cast.  The cast was the wrong answer to a correct warning: the type was
wrong, not the expression.  It is gone.

The colour-RAM arithmetic still computes in 32 bits, because
COLOUR_RAM_ADDRESS is $1F800 and so is a long literal, which promotes
the expression around it.

MONITOR drops 78 bytes -- 16-bit arithmetic where there had been 32-bit.
The other six targets are unchanged.

+++ src/fdisk_hal_mega65.c
@@ -291,7 +291,7 @@ void sdcard_writesector(const uint32_t sector_number, uint8_t is_multi) {
             if (i != 512) {
                 // VErify error has occurred
                 //	write_line("Verify error for sector $$$$$$$$",0);
-                screen_hex((unsigned int)(screen_line_address - 80 + 24), sector_number);
+                screen_hex(screen_line_address - 80 + 24, sector_number);
             } else {
                 //      write_line("Wrote sector $$$$$$$$, result=$$",2);
                 //      screen_hex(screen_line_address-80+2+14,sector_number);
